requirements.txt
aiogram<3.0
environs~=8.0.0
sqlalchemy~=1.3.24
asyncpg
DateTime~=4.3
tortoise-orm
aioredis
